requirements.txt
altair==6.1.0
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.10.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.1.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==48.0.0
distro==1.9.0
Django==5.2.7
dnspython==2.8.0
fastapi==0.115.8
gitdb==4.0.12
GitPython==3.1.50
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.196.0
google-auth==2.52.0
google-auth-httplib2==0.4.0
google-genai==2.0.0
googleapis-common-protos==1.75.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.14.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
motor==3.7.0
narwhals==2.20.0
numpy==2.2.2
openai==1.61.1
packaging==26.2
pandas==3.0.2
pillow==11.1.0
proto-plus==1.28.0
protobuf==5.29.6
pyarrow==24.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.4
pydantic-settings==2.7.1
pydantic_core==2.46.4
pydeck==0.9.2
pymongo==4.17.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.20
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
streamlit==1.57.0
stripe==11.6.0
tenacity==9.1.4
toml==0.10.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.7.0
uvicorn==0.34.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
wheel==0.45.1
